fix(build): move test files out of pages directory
